Reorganized and made code non-AU specific
